+++ mitra_models/migrations/V0092__media_attachment_add_url.sql
@@ -1,0 +1,1 @@
+ALTER TABLE media_attachment ADD COLUMN url VARCHAR(2000);