+++ mitra_models/src/media/types.rs
@@ -8,14 +8,23 @@ pub struct MediaInfo {
     pub file_name: String,
     pub file_size: usize,
     pub media_type: String,
+    pub url: Option<String>,
 }
 
-impl From<FileInfo> for MediaInfo {
-    fn from(file_info: FileInfo) -> Self {
+impl MediaInfo {
+    pub fn local(file_info: FileInfo) -> Self {
         Self {
             file_name: file_info.name,
             file_size: file_info.size,
             media_type: file_info.media_type,
+            url: None,
+        }
+    }
+
+    pub fn remote(file_info: FileInfo, url: String) -> Self {
+        Self {
+            url: Some(url),
+            ..Self::local(file_info)
         }
     }
 }