+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - De-duplicate media attachments by URL.
 - Use primary gateway address for filtering messages from portable actors.
 - Ignore images in `icon` field if object type is not `Video`.
+- Store URLs of remote media.
 
 ### Fixed